[Serialization] Pull swiftdoc serialization out to a separate file
No functionality change.
